Renamed files for `JoltCustomMotionShape`
