+++ src/spaces/jolt_motion_filter_3d.cpp
@@ -3,7 +3,7 @@
 #include "objects/jolt_body_impl_3d.hpp"
 #include "objects/jolt_object_impl_3d.hpp"
 #include "servers/jolt_physics_server_3d.hpp"
-#include "shapes/jolt_motion_shape.hpp"
+#include "shapes/jolt_custom_motion_shape.hpp"
 #include "shapes/jolt_shape_3d.hpp"
 #include "shapes/jolt_shape_type.hpp"
 #include "spaces/jolt_broad_phase_layer.hpp"
