+++ src/spaces/jolt_physics_direct_space_state_3d.cpp
@@ -5,7 +5,7 @@
 #include "objects/jolt_object_impl_3d.hpp"
 #include "servers/jolt_physics_server_3d.hpp"
 #include "servers/jolt_project_settings.hpp"
-#include "shapes/jolt_motion_shape.hpp"
+#include "shapes/jolt_custom_motion_shape.hpp"
 #include "shapes/jolt_shape_3d.hpp"
 #include "spaces/jolt_motion_filter_3d.hpp"
 #include "spaces/jolt_query_collectors.hpp"
